==> maps/1_overlap_between_maps_ancs.ipynb <==
# 1) Overlap Between Maps - ANCs

For a new redistricting map, show what old ANCs overlap with each new ANC.

Make sure rtree is installed in your local environment. See this notebook for your previous struggles: https://github.com/devinbrady/dc-parking/blob/main/Spatial-Join.ipynb

import pandas as pd
import geopandas as gpd
pd.set_option('display.float_format', lambda x: '%.2f' % x)

# Use this projection for measuring distances in meters
# https://octo.dc.gov/page/coordinate-system-standards
dc_crs = ('epsg', '26985')

# Use WGS84 for saving GeoJSON for display in QGIS
output_crs = ('epsg', '4326')

# Read in data
previous_map = gpd.read_file('anc-2012.geojson')
new_map = gpd.read_file('anc-2022.geojson')

# # Set column names on the new map
# new_map['smd_name'] = new_map['SMD_ID']
# new_map['smd_id'] = 'smd_2022_' + new_map['SMD_ID']
# new_map.rename(columns={'ANC_ID': 'anc_id'}, inplace=True)

# # We only need these columns going forward
# map_columns = ['smd_id', 'smd_name', 'anc_id', 'geometry']

# # Save out the new map for later steps
# # new_map[map_columns].to_file('to-mapbox-2022-smd-data-overlap.geojson', driver='GeoJSON')

# Convert to local coordinate reference system
previous_map = previous_map.to_crs(dc_crs)
new_map = new_map.to_crs(dc_crs)

previous_map['district_area'] = previous_map.geometry.area
new_map['district_area'] = new_map.geometry.area

def calculate_overlap(map_a, map_b, suffix_a, suffix_b, id_field):

    overlap = map_a[overlap_columns].overlay(
        map_b[overlap_columns], how='identity', keep_geom_type=False
    )
    
    overlap.rename(columns={
        id_field + '_1': id_field + suffix_a
        , id_field + '_2': id_field + suffix_b
        , 'district_area_1': 'district_area' + suffix_a
        , 'district_area_2': 'district_area' + suffix_b
    }, inplace=True)

    overlap['overlap_area'] = overlap.geometry.area
    overlap['overlap_perc'] = overlap['overlap_area'] / overlap['district_area' + suffix_a]
    
    overlap['district_rank'] = overlap.groupby(id_field + suffix_a).overlap_perc.rank(ascending=False)
    overlap.sort_values(by=[id_field + suffix_a, 'overlap_area'], ascending=[True, False], inplace=True)
    
    overlap.drop('geometry', axis=1, inplace=True)
    
    return overlap

# Calculate the overlap
overlap_columns = ['anc_id', 'district_area', 'geometry']

# Backwards - how much of each new district is made up of the old districts? 
# Use this for the frontend list for each new district. 
overlap_backwards = calculate_overlap(
    new_map[overlap_columns], previous_map[overlap_columns], '_2022', '_2012', 'anc_id')

# Forwards - how much of each old district went into this new district? 
# Use this for colors for continuity
overlap_forwards = calculate_overlap(
    previous_map[overlap_columns], new_map[overlap_columns], '_2012', '_2022', 'anc_id')

# Save results
columns_for_csv = [c for c in overlap_backwards.columns if c != 'geometry']
overlap_backwards[columns_for_csv].to_csv('overlap_backwards_anc.csv', index=False)
overlap_forwards[columns_for_csv].to_csv('overlap_forwards_anc.csv', index=False)

# Should be 46
(overlap_backwards.district_rank == 1).sum()

# Should be 40
(overlap_forwards.district_rank == 1).sum()

## Examples

Every old district that is in this new district:

overlap_forwards[overlap_forwards.anc_id_2022 == 'anc_1C_2022']

Every new district that is made up of parts of this old district:

overlap_forwards[overlap_forwards.anc_id_2012 == 'anc_1C']

Every new district that this old district went to make: 

overlap_backwards[overlap_backwards.anc_id_2022 == 'anc_1C_2022']

overlap_backwards[overlap_backwards.anc_id_2012 == 'anc_1C']

overlap_forwards.overlap_perc.sum()

overlap_backwards.overlap_perc.sum()
