==> maps/2_assign_map_colors.ipynb <==
# 2) Assign Map Colors

For a new map after redistricing, try to make the colors match when switching between map versions.

Goal: If 1C01 is light blue on the 2012 map, then the same area should also be light blue on the 2022 map. The district number might have changed, though! 

import pandas as pd
import geopandas as gpd

districts = pd.read_csv('../data/districts.csv')
overlap = pd.read_csv('overlap_forwards.csv')

# Should be 345
map_2022 = gpd.read_file('smd-2022-overlap.geojson')
len(map_2022)

# Should be 296 - the new district that each old district most overlaps
overlap_top_forwards = overlap[overlap.district_rank == 1].copy()
len(overlap_top_forwards)

There are situations where two old districts overlap the most with the same new district. For instance, the old 1C02 and 1C04 both overlap the most with the new 1C04. In these cases, we want to "connect" the old district to the new district where both share the most overlap area. 

overlap_top_forwards['overlap_area_rank'] = (
    overlap_top_forwards.groupby('smd_id_2022').overlap_area.rank(ascending=False)
)
overlap_top_forwards[overlap_top_forwards['smd_id_2022'] == 'smd_2022_1C04']

overlap_top_match = overlap_top_forwards[overlap_top_forwards.overlap_area_rank == 1].copy()

# Left join the 2022 map to the 2012 district overlap
district_overlap = pd.merge(
    districts.loc[districts.redistricting_year == 2012, ['smd_id', 'map_color_id']]
    , overlap_top_match
    , how='left'
    , left_on='smd_id'
    , right_on='smd_id_2012'
)

# Drop this duplicate column to avoid confusion and collision
district_overlap.drop(columns=['smd_id'], inplace=True)

# Should be 296
len(district_overlap)

map_2022_colors = map_2022.merge(
    district_overlap, how='left', left_on='smd_id', right_on='smd_id_2022'
)

# Should be 345
len(map_2022_colors)

map_2022_colors['map_color_id'] = map_2022_colors['map_color_id'].fillna(13)

# Make sure this column in an integer. Because it was previously NULL, pandas set it a decimal
map_2022_colors['map_color_id'] = map_2022_colors['map_color_id'].astype('int64')

map_2022_colors['qgis_label'] = '2022_' + map_2022_colors['smd_name']

map_2022_colors.head()

map_2022_colors.groupby('map_color_id').size()

# Should be 345
len(map_2022_colors)

map_columns = ['smd_id', 'smd_name', 'anc_id', 'geometry']

map_2022_colors[map_columns + ['map_color_id', 'qgis_label']].to_file(
    'smd-2022-colors.geojson', driver='GeoJSON')

# no_geo = [c for c in map_2022_overlap.columns if c != 'geometry']
# map_2022_overlap[no_geo].sort_values(by='smd_id').to_clipboard()
