==> maps/5_new_districts_2022.ipynb <==
# 5) New Districts 2022

Format district data for adding to existing `districts` table.

import pandas as pd
import geopandas as gpd

districts = pd.read_csv('../data/districts.csv')
map_2022 = gpd.read_file('smd-2022-preprocessed.geojson')
label_points = pd.read_csv('label-points-2022.csv')
new_map = pd.merge(map_2022, label_points[['smd_id', 'lon', 'lat']], how='inner', on='smd_id')

# Should be 345
len(new_map)

comp = pd.merge(new_map, districts, how='left', on='smd_id')

# New districts in 2022
# comp[comp.anc_id_y.isnull()].smd_id.tolist()

# new_map.groupby('anc_id').size()

# Specify the ward each SMD is in
new_map['ward'] = new_map['anc_id'].str[0]

# Manually set the 3/4G wards
new_map.loc[new_map.smd_id.isin([
    'smd_2022_3/4G05'
    , 'smd_2022_3/4G06'
    , 'smd_2022_3/4G07'
]), 'ward'] = 3

new_map.loc[new_map.smd_id.isin([
    'smd_2022_3/4G01'
    , 'smd_2022_3/4G02'
    , 'smd_2022_3/4G03'
    , 'smd_2022_3/4G04'
]), 'ward'] = 4

new_map['centroid_lon'] = new_map['lon']
new_map['centroid_lat'] = new_map['lat']
new_map['neighbor_smds'] = new_map['neighbors']
new_map['redistricting_cycle'] = 2022
new_map['smd_name'] = new_map['smd_id'].str.replace('smd_2022_', '')

output_columns = [
    'smd_id'
    , 'redistricting_cycle'
    , 'smd_name'
    , 'anc_id'
    , 'ward'
    , 'centroid_lon'
    , 'centroid_lat'
    , 'area'
    , 'map_color_id'
    , 'neighbor_smds'
]

new_map[output_columns]

# new_map[output_columns].to_clipboard(index=False)

## List of SMDs for candidate declaration form

form_values = 'SMD ' + districts[districts.redistricting_year == 2022].sort_values(by='sort_order').smd_name
# form_values.to_clipboard(index=False)

form_values

## List of ANCs

ancs_2012 = sorted(districts[districts.redistricting_year == 2012].anc_id.unique())
ancs_2022 = sorted(districts[districts.redistricting_year == 2022].anc_id.unique())

# New ANCs
[a for a in ancs_2022 if a not in ancs_2012]

[a for a in ancs_2012 if a not in ancs_2022]
