==> maps/6_list_of_historical_overlapping_districts.ipynb <==
# 6) List of historical overlapping districts

For every SMD past and present, generate a list of SMDs that overlapped with it.

In other words, turn the overlap CSVs (one row for every overlap) into a list for every district.

* Backwards - how much of each new district is made up of the old districts? 
* Forwards - how much of each old district went into this new district? 

import pandas as pd

districts = pd.read_csv('../data/districts.csv')
overlap_backwards = pd.read_csv('overlap_backwards.csv')
overlap_forwards = pd.read_csv('overlap_forwards.csv')

Confirm that the overlap percentages add up to roughly one

all(abs(overlap_backwards.groupby('smd_id_2022').overlap_perc.sum() - 1) < 0.0001)

all(abs(overlap_forwards.groupby('smd_id_2012').overlap_perc.sum() - 1) < 0.0001)

overlap_backwards['valid'] = True
overlap_forwards['valid'] = True

# The overlap area must by at least 10 meters square
overlap_backwards.loc[overlap_backwards.overlap_area < 10, 'valid'] = False
overlap_forwards.loc[overlap_forwards.overlap_area < 10, 'valid'] = False

overlap_backwards.loc[overlap_backwards.smd_id_2012.isnull(), 'valid'] = False
overlap_backwards.loc[overlap_backwards.smd_id_2022.isnull(), 'valid'] = False

overlap_forwards.loc[overlap_forwards.smd_id_2012.isnull(), 'valid'] = False
overlap_forwards.loc[overlap_forwards.smd_id_2022.isnull(), 'valid'] = False

# Manually remove some overlaps

# Per Corey Holman, this overlap is just a shapefile artifact along a bridge
overlap_backwards.loc[
    (overlap_backwards.smd_id_2012 == 'smd_6B06') & (overlap_backwards.smd_id_2022 == 'smd_2022_8F01')
    , 'valid'
] = False

overlap_forwards.loc[
    (overlap_forwards.smd_id_2012 == 'smd_6B06') & (overlap_forwards.smd_id_2022 == 'smd_2022_8F01')
    , 'valid'
] = False

# Limit to only the valid overlaps
overlap_backwards = overlap_backwards[overlap_backwards.valid].copy()
overlap_forwards = overlap_forwards[overlap_forwards.valid].copy()

# Sort by overlap percentage, descending within SMDs
overlap_forwards.sort_values(by=['smd_id_2012', 'district_rank'])

# of_list = overlap_forwards.groupby('smd_id_2022').smd_id_2012.apply(lambda x: ', '.join(sorted(list(x))))

# ob_list = overlap_backwards.groupby('smd_id_2012').smd_id_2022.apply(lambda x: ', '.join(sorted(list(x))))

# df = pd.DataFrame(pd.concat([of_list, ob_list]), columns=['overlap_smds'])
# df.index.name = 'smd_id'
# df = df.reset_index()

# df['redistricting_year'] = df.smd_id.apply(lambda x: 2022 if '_2022_' in x else 2012)

# Copy this to OpenANC Source
# df.sort_values(by=['redistricting_year', 'smd_id']).overlap_smds.to_clipboard(index=False)

# Convert the overlap percentages into strings so that they can be nicely concatenated
overlap_backwards['overlap_perc_str'] = overlap_backwards.overlap_perc.apply(lambda x: f'{x:.4f}')
overlap_forwards['overlap_perc_str'] = overlap_forwards.overlap_perc.apply(lambda x: f'{x:.4f}')

overlap_backwards.groupby('smd_id_2022').agg(
    overlap_smds=('smd_id_2012', lambda x: ', '.join(list(x)))
    , overlap_percentage=('overlap_perc_str', lambda x: ', '.join(list(x)))
).to_clipboard()

overlap_forwards.groupby('smd_id_2012').agg(
    overlap_smds=('smd_id_2022', lambda x: ', '.join(list(x)))
    , overlap_percentage=('overlap_perc_str', lambda x: ', '.join(list(x)))
).to_clipboard()
